Fix call to gpu_density
gpu_density() now recieves the bandwidth parameter of the Gaussian
kernel rather than a handle of the density kernel itself.

Compute the values for bandwidth parameter such that pair of subsequences
that are 3*noise_variance apart will have a density kernel value of 0.1.
Since the density evaluations from multiple subsequence pairs are summed for
a fixed subsequence, the intuition is that when a pair is apart by the maximum
noise then it will contribute only 10% to that sum. The variance of the noise
is computed using "reasonable" values of SNR for an EEG signal and the fact that
the signals are z-normalized (have unit variance).