Allow multiple density estimates as input
gpu_density() computes the density estimate for multiple values of
the bandwidth parameter of the Gaussian kernel, which leads to
multiple density estimates that gpu_qsmp() can use to compute different
qsmp profiles and indices.

Also, set QSMP to numpy.nan for the subsquences that fall in the
splice of segments, and to numpy.inf if there is no trivial nearest
neighbor that increases the density.