Fix computation of profile and index
The distance was being modified inside the for loop that iterates over the
densities. This is now fixed.

Additional code has been added to periodically save the computations of profile
and index. The period is defined in the config module with the variable
QSMP_CHECKPOINT_PERIOD. This is useful if you are running your job with SLURM
and it gets preempted.