Create find_modes_no_exclusion_zone
Nearest-neighbor (NN) distance and index are now computed without using a
exclusion zone. This makes the tree highly fragmented and more expensive
to traverse using the method in find_modes().

Here we order the density, NN-distance (`profile`), and NN-index
(`neighbor`) in descending order of the density. Then, we pick as modes
the points where the profile is bigger than the distance threshold
(`maxdist`).