Add code to compute autocorr and fwhm to scale distance
Slow waveforms tend to have smaller distances and thus a much bigger
density. Also, the full width half maximum (FWHM) of the (global
maximum of the) autocorrelation of the slow waveforms is small.
Faster waveforms behave in the opposite way. From this observation,
we want to test if scaling the distance between waveforms by their FWHM
could boost the density of faster waveforms. This can be thought as a
"whitening" of the distance (or density) so that waveforms are more
uniformly processed, independent of their rate of change.