Add new functions to cluster subsequences
These new functions are inspired by the scikit-learn implementation of
Quick Shift. It's a much faster approach to the other functions that
uses the tree data structure and the Node class. After clustering, we
lose the information about the original nearest neighbor of each
subsequence, which is OK as that information is irrelevant once we
clustered and we can recompute the distances of each subsequence
in the cluster to their root.