Fix computation of FWHM for concatenated time series
For a time series that is the result of `n_seg` concatenated segments,
the autocorrelation needs to be computed on each segment separately.
Also, the computation of FWHM now consider the case where the
autocorrelation is monotonically increasing from the start to the global
maxima, i.e., there are no side lobes, only one broad main lobe centered
at the global maximizer.