Refactor: Compute shift-invariant distance (on CPU)
We want to incorporate shift-invariance in the computation of
NN-distance and NN-index. This refactoring uses move_argmin from the
bottleneck library to implement a rolling argmin on the distance profile
of each subsequence. Although move_argmin is fast, the combination of
this and an numpy.argmin operation takes 100 times the time of computing
the distance profile with the GPU kernel. So, we might need to move this
operations to GPU.